dataset-ingestor/requirements.txt
pandas
pyarrow
google-cloud-storage
google-cloud-bigquery
kaggle
dotenv